feat/marked 15 - fix/demo structure and validations - fix/test and release workflows
