+++ .github/workflows/release-demo.yml
@@ -24,7 +24,8 @@ jobs:
   build:
     if: |
       github.ref == 'refs/heads/main' &&
-      github.event.repository.fork == false
+      github.event.repository.fork == false &&
+      github.actor != 'dependabot[bot]'
     runs-on: ubuntu-latest
     timeout-minutes: 1
     environment:
@@ -65,10 +66,6 @@ jobs:
           if-no-files-found: error
 
   release:
-    if: |
-      github.ref == 'refs/heads/main' &&
-      github.event.repository.fork == false &&
-      github.actor != 'dependabot[bot]'
     runs-on: ubuntu-latest
     timeout-minutes: 1
     environment:
