+++ .github/workflows/release-library.yml
@@ -17,7 +17,7 @@ on:
 
 concurrency:
   group: 'release-library'
-  cancel-in-progress: true
+  cancel-in-progress: false
 
 env:
   ARTIFACT_NAME: 'build-lib-artifact'
@@ -26,7 +26,8 @@ jobs:
   build:
     if: |
       github.ref == 'refs/heads/main' &&
-      github.event.repository.fork == false
+      github.event.repository.fork == false &&
+      github.actor != 'dependabot[bot]'
     runs-on: ubuntu-latest
     timeout-minutes: 2
     environment: FSI_DEP_NodeJs
@@ -82,11 +83,7 @@ jobs:
             echo "TAG=latest" >> $GITHUB_OUTPUT
           fi
 
-      - name: Release to NPM registry 🚀
-        if: |
-          github.ref == 'refs/heads/main' &&
-          github.event.repository.fork == false &&
-          github.actor != 'dependabot[bot]'
+      - name: Release to NPM registry 🚀 - ${{ steps.get_version.outputs.VERSION }}
         env:
           NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }}
         run: |
@@ -101,11 +98,7 @@ jobs:
           scope: ${{ vars.NPM_ORG }}
           always-auth: true
 
-      - name: Release to GitHub NPM registry 🚀
-        if: |
-          github.ref == 'refs/heads/main' &&
-          github.event.repository.fork == false &&
-          github.actor != 'dependabot[bot]'
+      - name: Release to GitHub NPM registry 🚀 - ${{ steps.get_version.outputs.VERSION }}
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           NODE_AUTH_TOKEN: ${{ secrets.NPM_GH_TOKEN }}
